nodo-x402/requirements.txt
# NODO x402 - Solana Payment Protocol for AI Analysis
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Solana
solana>=0.32.0
solders>=0.20.0
anchorpy>=0.18.0
spl-token>=0.3.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# AI APIs
anthropic>=0.18.0
openai>=1.10.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
redis>=5.0.0
alembic>=1.13.0

# Caching & Rate Limiting
aiocache>=0.12.0
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0

# Crypto utilities
base58>=2.1.0
nacl>=1.5.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0


